Explode the checkout module.
Make it a real package, and then split it into reasonable modules.
