+++ dss/api/bundles/checkout.py
@@ -4,7 +4,8 @@ from flask import jsonify
 
 from dss import dss_handler, Replica
 from dss.error import DSSException
-from dss.storage.checkout import BundleNotFoundError, CheckoutStatus, start_bundle_checkout
+from dss.storage.checkout import BundleNotFoundError
+from dss.storage.checkout.bundle import get_bundle_checkout_status, start_bundle_checkout
 
 
 @dss_handler
@@ -31,7 +32,7 @@ def get(replica: str, checkout_job_id: str):
     assert replica is not None
     _replica = Replica[replica]
     try:
-        response = CheckoutStatus.get_bundle_checkout_status(checkout_job_id, _replica, _replica.checkout_bucket)
+        response = get_bundle_checkout_status(checkout_job_id, _replica, _replica.checkout_bucket)
     except BlobNotFoundError:
         raise DSSException(requests.codes.not_found, "not_found", "Cannot find checkout!")
     return response, requests.codes.ok
