+++ dss/api/files.py
@@ -13,7 +13,8 @@ from flask import jsonify, make_response, redirect, request
 
 from dss import DSSException, dss_handler, stepfunctions
 from dss.config import Config, Replica
-from dss.storage.checkout import CheckoutTokenKeys, get_dst_key, start_file_checkout
+from dss.storage.checkout import CheckoutTokenKeys
+from dss.storage.checkout.file import get_dst_key, start_file_checkout
 from dss.storage.files import write_file_metadata
 from dss.storage.hcablobstore import FileMetadata, HCABlobStore, compose_blob_key
 from dss.stepfunctions import gscopyclient, s3copyclient
